extract Select to a reusable dumb component
