+++ applications/launchpad_v2/src/styles/WithTheme.tsx
@@ -1,0 +1,8 @@
+import { useTheme } from './useTheme'
+
+export const WithTheme = (Component) => function TT(props) {
+  const theme = useTheme()
+
+  return <Component {...props} theme={theme} />
+}
+
