+++ applications/launchpad_v2/src/styles/index.ts
@@ -1,1 +1,1 @@
-export * from './useTheme'
+export * from './WithTheme'
